Cleaned up callbacks controller a little
Removed unused variables and method params.
Since we have 'name', pass that instead of the whole hash to methods.

'auth' is repeatedly overwritten - removed some of that.

This still looks suspect:

    unless user.authorizations.find_by_provider("twitter")
      auth = user.authorizations.build(provider: "twitter", uid: uid)
      user.authorizations << auth
    end

Since it overwrites auth then builds an authorization then adds that to
a list that I assume build would add it to anyway.

Then later we call:
    auth.update_attributes(auth_attributes(auth_hash))

...which would presumably replace the auth we just added? I don't know
how this is all supposed to work.

I've not tested this code, just attempted to refactor it without
changing the logic too much.